requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.2
aiosignal==1.3.1
alembic==1.14.0
annotated-types==0.7.0
anyio==3.7.1
asgiref==3.8.1
attrs==24.2.0
backoff==2.2.1
bcrypt==4.0.1
beautifulsoup4==4.12.3
Brotli==1.1.0
build==1.2.2.post1
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
chroma-hnswlib==0.7.3
chromadb==0.4.24
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
crewai==0.22.5
crewai-tools==0.0.15
cryptography==43.0.3
dataclasses-json==0.6.7
decorator==5.1.1
Deprecated==1.2.15
deprecation==2.1.0
distro==1.9.0
dnspython==2.7.0
docstring-parser==0.15
durationpy==0.9
ecdsa==0.19.0
email-validator==2.1.0.post1
embedchain==0.1.102
et_xmlfile==2.0.0
exa_py==1.6.0
faiss-cpu==1.9.0
fastapi==0.104.1
filelock==3.16.1
flatbuffers==24.3.25
frozenlist==1.5.0
fsspec==2024.10.0
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.4
google-api-core==2.23.0
google-api-python-client==2.153.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.72.0
google-cloud-bigquery==3.27.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.13.1
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-generativeai==0.5.4
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
gptcache==0.1.44
greenlet==3.1.1
groq==0.12.0
grpc-google-iam-v1==0.13.1
grpcio==1.67.1
grpcio-status==1.62.3
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.27.2
huggingface-hub==0.26.2
humanfriendly==10.0
idna==3.10
importlib_metadata==8.4.0
importlib_resources==6.4.5
iniconfig==2.0.0
instructor==0.5.2
jiter==0.7.1
jsonpatch==1.33
jsonpointer==3.0.0
kubernetes==31.0.0
lancedb==0.5.7
langchain==0.1.20
langchain-community==0.0.38
langchain-core==0.1.53
langchain-google-genai==1.0.4
langchain-groq==0.1.5
langchain-openai==0.0.5
langchain-postgres==0.0.9
langchain-text-splitters==0.0.2
langsmith==0.1.81
lxml==5.3.0
Mako==1.3.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
mdurl==0.1.2
mmh3==5.0.1
monotonic==1.6
mpmath==1.3.0
multidict==6.1.0
mutagen==1.47.0
mypy-extensions==1.0.0
nodeenv==1.9.1
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.20.0
openai==1.54.4
openpyxl==3.1.5
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-exporter-otlp-proto-http==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-asgi==0.48b0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opentelemetry-util-http==0.48b0
orjson==3.10.11
outcome==1.3.0.post0
overrides==7.7.0
packaging==23.2
pandas==2.2.3
passlib==1.7.4
pgvector==0.2.5
pluggy==1.5.0
posthog==3.7.0
propcache==0.2.0
proto-plus==1.25.0
protobuf==4.25.5
psycopg==3.2.3
psycopg-pool==3.2.4
psycopg2-binary==2.9.9
pulsar-client==3.5.0
py==1.11.0
pyarrow==18.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodomex==3.21.0
pydantic==2.9.2
pydantic_core==2.23.4
PyGithub==1.59.1
Pygments==2.18.0
PyJWT==2.9.0
pylance==0.9.18
PyNaCl==1.5.0
pyparsing==3.2.0
pypdf==3.17.4
PyPDF2==3.0.1
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
pyright==1.1.389
pysbd==0.3.4
PySocks==1.7.1
pytest==8.3.3
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytube==15.0.0
pytz==2024.2
PyYAML==6.0.2
ratelimiter==1.2.0.post0
regex==2023.12.25
requests==2.32.3
requests-oauthlib==2.0.0
retry==0.9.2
rich==13.9.4
rsa==4.9
schema==0.7.7
selenium==4.26.1
semver==3.0.2
setuptools==75.5.0
shapely==2.0.6
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.36
starlette==0.27.0
sympy==1.13.3
tenacity==8.5.0
tiktoken==0.5.2
tokenizers==0.20.3
tqdm==4.67.0
trio==0.27.0
trio-websocket==0.11.1
typer==0.9.4
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.24.0
watchfiles==0.24.0
websocket-client==1.8.0
websockets==14.1
wrapt==1.16.0
wsproto==1.2.0
yarl==1.17.1
youtube-transcript-api==0.6.2
yt-dlp==2023.12.30
zipp==3.21.0